Conflicts:
	Firmware/environment/PEGMARev0.uvproj
	Firmware/source/main.c

+++ Firmware/header/util.h
@@ -18,7 +18,7 @@ void Util_reverseBytes(uint8 *pToRev, uint8 numBytes);
 void   Util_uint16ToASCII(uint16 inVal, char *outChars);
 int8   Util_compareMemory(uint8* pLeft, uint8* pRight, uint16 numBytes);
 void   Util_fillMemory(void *pDst, uint32 numBytes, uint8 fillVal);
-void   Util_copyMemory(uint8* pSrc, uint8* pDst, uint16 numBytes);
+void   Util_copyMemory(uint8* pSrc, uint8* pDst, uint32 numBytes);
 void   Util_spinWait(volatile uint32 spinVal);
 void   Util_spinDelay(uint32 microSecondDelay);
 uint32 Util_reflect(uint32 inWord, uint32 nbrToReflect);
